Split processUsage types into separate file
