+++ src/app/api/dev/consume-credits/route.ts
@@ -3,9 +3,10 @@ import { NextResponse } from 'next/server';
 import { getUserFromAuth } from '@/lib/user.server';
 import { findUserById } from '@/lib/user';
 import { forceImmediateExpirationRecomputation } from '@/lib/balanceCache';
-import { countAndStoreUsage, type MicrodollarUsageContext } from '@/lib/processUsage';
+import { countAndStoreUsage } from '@/lib/processUsage';
 import { captureException } from '@sentry/nextjs';
 import { getFraudDetectionHeaders } from '@/lib/utils';
+import type { MicrodollarUsageContext } from '@/lib/processUsage.types';
 
 export async function POST(request: NextRequest): Promise<NextResponse> {
   // Check if we're in development mode
